big update vacance noel

+++ commands/clear.js
@@ -3,6 +3,7 @@ exports.run = async (client, message, args) => {
     if (!message.member.hasPermission("MANAGE_MESSAGES")) return message.channel.send("Pas les perms.");
     if (isNaN(nombre)) return message.channel.send('pas de nombre');
     var n = parseInt(nombre)+1;
+    if (n > 100) return message.channel.send('100 max')
     message.channel.bulkDelete(n, true)
     message.channel.send(nombre+" messages supprimés");
     setTimeout(() => {
